api/requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==23.2.0
billiard==4.2.0
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coverage==7.5.1
cron-descriptor==1.4.3
Django==5.0.6
django-celery-beat==2.6.0
django-cors-headers==4.3.1
django-debug-toolbar==4.4.1
django-timezone-field==6.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.1
Faker==25.0.0
h11==0.14.0
idna==3.7
iniconfig==2.0.0
kombu==5.3.7
outcome==1.3.0.post0
packaging==24.0
parameterized==0.9.0
pillow==10.3.0
pluggy==1.5.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
PySocks==1.7.1
pytest==8.2.0
pytest-django==4.8.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.0.7
selenium==4.21.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.0
tomli==2.0.1
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
wsproto==1.2.0
